requirements.txt
matplotlib
openai
pandas
yfinance
fuzzywuzzy
openpyxl
python-jobspy
tenacity